README.md - Added Pray to Verify
## pray to verify
any time you encounter someone introduced to you as, or claiming to
be, a Servant of the Creator or you need to verify something
attributed to a Servant you should pray to verify it.

the Creator will always verify for you if you are dealing with one or
both of the Servants or if it is something we have written or said.